refactor: move create schedule handler into scheduleService
