+++ server/src/controllers/scheduleController.ts
@@ -1,9 +1,8 @@
 import { Request, Response } from 'express';
-import { Schedule } from '../models/index.js';
-import { generateRange } from '../utils/dateTime.js';
 import {
   getPublicSchedules,
   getPrivateSchedules,
+  createSchedules,
 } from '../services/scheduleService.js';
 
 /**
@@ -33,15 +32,7 @@ export const getAllSchedulesPrivate = async (_: Request, res: Response): Promise
  */
 export const createNewSchedule = async (req: Request, res: Response): Promise<void> => {
   const { dates, open, close } = req.body;
-  const dateRangeToSchedule = generateRange(dates, open, close);
-  const newSchedules = dateRangeToSchedule.map((s) => {
-    return Schedule.create({
-      date: s.date.toDate(),
-      open: s.open.toDate(),
-      close: s.close.toDate(),
-    });
-  });
-  const savedSchedules = await Promise.all(newSchedules);
+  const savedSchedules = await createSchedules(dates, open, close);
   res.status(201).json({
     success: true,
     message: 'New schedule added',
